+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1. [15836](https://github.com/influxdata/influxdb/pull/16077): Add stacked line layer option to graphs
 1. [16094](https://github.com/influxdata/influxdb/pull/16094): Annotate log messages with trace ID, if available
+1. [16187](https://github.com/influxdata/influxdb/pull/16187): Bucket create to accept an org name flag
 
 ### Bug Fixes